03-ReservasHotel/requirements.txt
FastAPI
SQLAlchemy
python-dotenv
uvicorn
pydantic
pydantic[email]
psycopg[binary]
alembic